refactor: move Crc32C to internals
