+++ clients/src/main/java/org/apache/kafka/common/record/internal/DefaultRecordBatch.java
@@ -24,7 +24,7 @@ import org.apache.kafka.common.header.Header;
 import org.apache.kafka.common.record.TimestampType;
 import org.apache.kafka.common.utils.BufferSupplier;
 import org.apache.kafka.common.utils.ByteBufferOutputStream;
-import org.apache.kafka.common.utils.Crc32C;
+import org.apache.kafka.common.utils.internals.Crc32C;
 import org.apache.kafka.common.utils.internals.ByteUtils;
 import org.apache.kafka.common.utils.internals.CloseableIterator;
