+++ storage/src/main/java/org/apache/kafka/storage/internals/log/ProducerStateManager.java
@@ -20,7 +20,7 @@ import org.apache.kafka.common.TopicPartition;
 import org.apache.kafka.common.protocol.ByteBufferAccessor;
 import org.apache.kafka.common.protocol.MessageUtil;
 import org.apache.kafka.common.record.internal.RecordBatch;
-import org.apache.kafka.common.utils.Crc32C;
+import org.apache.kafka.common.utils.internals.Crc32C;
 import org.apache.kafka.common.utils.LogContext;
 import org.apache.kafka.common.utils.Time;
 import org.apache.kafka.common.utils.Utils;
